Move types into their modules
