+++ src/modules/api/BittrexApi.ts
@@ -14,7 +14,7 @@ import {
   OrderDirection,
   OrderType,
   TimeInForce
-} from '../../types';
+} from './types';
 import fetch, { Headers, Response } from 'node-fetch';
 
 interface BittrexClientOptions {
@@ -229,18 +229,15 @@ export default class BittrexApi {
    * @param BTC - current market USDT price of BTC
    */
   report = async (USDT: number, BTC: number) => {
-    await fetch(
-      'https://YOUR_URL_HERE',
-      {
-        headers: {
-          'Content-Type': 'application/json'
-        },
-        method: 'POST',
-        body: JSON.stringify({
-          totalBalance: Math.floor(USDT),
-          currentMarketBTC: Math.floor(BTC)
-        })
-      }
-    );
+    await fetch('https://YOUR_URL_HERE', {
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      method: 'POST',
+      body: JSON.stringify({
+        totalBalance: Math.floor(USDT),
+        currentMarketBTC: Math.floor(BTC)
+      })
+    });
   };
 }
