+++ src/modules/api/types.ts
@@ -33,17 +33,6 @@ export interface Candle {
   quoteVolume: number;
 }
 
-/**
- * Only used for internal calculations for the react-stockcharts library
- * e.g. for ema
- */
-export interface CandleReactStockCharts
-  extends Omit<Candle, 'startsAt' | 'quoteVolume'> {
-  date: Date;
-  ema9: number;
-  ema26: number;
-}
-
 export interface CandleResponse
   extends Omit<
     Candle,
